// src/app/interfaces/group-info.ts
export interface GroupInfo {
  groupID: number;
  groupName: string;
  clientCode: string;
  companyName: string;
  location: string;
  address1: string;
  city: string;
  state: string;
  zip: string;
  privateCharges: number;
  defaultPlant: string;
  groupType: number;
  tiberID: number;
  logoAddress: string;
  logoHeight: number;
  logoWidth: number;
  country: string;
  contactName: string;
  contactPhone: string;
  contactEmail: string;
  isP44DispatchEnabled: number;
  defaultReadyBy: number;
  defaultPickupBy: number;
}